Move tests to respective invocation layers
